+++ raft/snapshot.go
@@ -63,7 +63,7 @@ func (pm *ProtocolManager) buildSnapshot() *SnapshotWithHostnames {
 
 	// Populate addresses
 	for i, rawRaftId := range append(pm.confState.Nodes, pm.confState.Learners...) {
-		raftId := uint64(rawRaftId)
+		raftId := rawRaftId
 
 		if raftId == pm.raftId {
 			snapshot.Addresses[i] = *pm.address
@@ -117,7 +117,7 @@ func (pm *ProtocolManager) triggerSnapshot(index uint64) {
 func confStateIdSet(confState raftpb.ConfState) mapset.Set {
 	set := mapset.NewSet()
 	for _, rawRaftId := range append(confState.Nodes, confState.Learners...) {
-		set.Add(uint64(rawRaftId))
+		set.Add(rawRaftId)
 	}
 	return set
 }
